chore: introduce private build of the CLI
