+++ cliv2-private/cmd/cliv2/main.go
@@ -1,0 +1,15 @@
+package main
+
+import (
+	"os"
+
+	"github.com/snyk/remy-cli-extension/pkg/remy"
+
+	"github.com/snyk/cli/cliv2/pkg/runner"
+)
+
+func main() {
+	os.Exit(runner.Run(
+		runner.WithAdditionalExtensions(remy.Init),
+	))
+}
