+++ cliv2/pkg/runner/errorhandling_test.go
@@ -1,4 +1,4 @@
-package main
+package runner
 
 import (
 	"errors"
