+++ cliv2/pkg/runner/exitcode.go
@@ -1,4 +1,4 @@
-package main
+package runner
 
 import (
 	"context"
