+++ cliv2/pkg/runner/exitcode_test.go
@@ -1,4 +1,4 @@
-package main
+package runner
 
 import (
 	"context"
