+++ cliv2/pkg/runner/logheaderfooter.go
@@ -1,7 +1,4 @@
-package main
-
-// !!! This import needs to be the first import, please do not change this !!!
-import _ "github.com/snyk/go-application-framework/pkg/networking/fips_enable"
+package runner
 
 import (
 	"crypto/sha256"
