+++ cliv2/pkg/runner/runner.go
@@ -1,0 +1,751 @@
+package runner
+
+// !!! This import needs to be the first import, please do not change this !!!
+import _ "github.com/snyk/go-application-framework/pkg/networking/fips_enable"
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"io"
+	"os"
+	"os/exec"
+	"strings"
+	"sync"
+	"time"
+
+	"github.com/google/uuid"
+	"github.com/rs/zerolog"
+	"github.com/snyk/cli-extension-agent-scan/pkg/agentscan"
+	"github.com/snyk/cli-extension-ai-bom/pkg/aibom"
+	"github.com/snyk/cli-extension-ai-redteam/pkg/redteam"
+	"github.com/snyk/cli-extension-dep-graph/pkg/depgraph"
+	"github.com/snyk/cli-extension-iac-rules/iacrules"
+	"github.com/snyk/cli-extension-iac/pkg/iac"
+	"github.com/snyk/cli-extension-os-flows/pkg/osflows"
+	"github.com/snyk/cli-extension-sbom/pkg/sbom"
+	"github.com/snyk/cli-extension-secrets/pkg/secrets"
+	"github.com/snyk/code-client-go/pkg/code"
+	"github.com/snyk/container-cli/pkg/container"
+	"github.com/snyk/error-catalog-golang-public/cli"
+	"github.com/spf13/cobra"
+	"github.com/spf13/pflag"
+
+	"github.com/snyk/go-application-framework/pkg/ui/consoleui"
+
+	"github.com/snyk/go-application-framework/pkg/analytics"
+	"github.com/snyk/go-application-framework/pkg/app"
+	"github.com/snyk/go-application-framework/pkg/configuration"
+	"github.com/snyk/go-application-framework/pkg/instrumentation"
+	"github.com/snyk/go-application-framework/pkg/logging"
+
+	"github.com/snyk/cli/cliv2/cmd/cliv2/behavior/legacy"
+	"github.com/snyk/cli/cliv2/internal/cliv2"
+	"github.com/snyk/cli/cliv2/internal/constants"
+
+	cliv2utils "github.com/snyk/cli/cliv2/internal/utils"
+
+	localworkflows "github.com/snyk/go-application-framework/pkg/local_workflows"
+	"github.com/snyk/go-application-framework/pkg/local_workflows/config_utils"
+	"github.com/snyk/go-application-framework/pkg/local_workflows/network_utils"
+
+	"github.com/snyk/go-httpauth/pkg/httpauth"
+	"github.com/snyk/snyk-iac-capture/pkg/capture"
+
+	workflows "github.com/snyk/go-application-framework/pkg/local_workflows/connectivity_check_extension"
+
+	ignoreworkflow "github.com/snyk/go-application-framework/pkg/local_workflows/ignore_workflow"
+	"github.com/snyk/go-application-framework/pkg/local_workflows/output_workflow"
+	"github.com/snyk/go-application-framework/pkg/networking"
+	"github.com/snyk/go-application-framework/pkg/networking/middleware"
+	"github.com/snyk/go-application-framework/pkg/runtimeinfo"
+	"github.com/snyk/go-application-framework/pkg/ui"
+	"github.com/snyk/go-application-framework/pkg/utils"
+	"github.com/snyk/go-application-framework/pkg/workflow"
+
+	snykls "github.com/snyk/snyk-ls/ls_extension"
+	"github.com/snyk/studio-mcp/pkg/mcp"
+
+	cli_errors "github.com/snyk/cli/cliv2/internal/errors"
+	"github.com/snyk/cli/cliv2/pkg/basic_workflows"
+)
+
+// Option configures the CLI runner.
+type Option func(*runnerConfig)
+
+type runnerConfig struct {
+	additionalExtensions []workflow.ExtensionInit
+}
+
+// WithAdditionalExtensions adds extension initializers to the CLI beyond the
+// core set.  This is the hook used by the enterprise/private entry point.
+func WithAdditionalExtensions(exts ...workflow.ExtensionInit) Option {
+	return func(c *runnerConfig) {
+		c.additionalExtensions = append(c.additionalExtensions, exts...)
+	}
+}
+
+// additionalExts is set by Run before calling MainWithErrorCode.
+var additionalExts []workflow.ExtensionInit
+
+// Run is the public entry point for the CLI. It accepts options to customise
+// behaviour — most notably WithAdditionalExtensions for injecting private
+// extension initializers.
+func Run(opts ...Option) int {
+	cfg := &runnerConfig{}
+	for _, opt := range opts {
+		opt(cfg)
+	}
+	additionalExts = cfg.additionalExtensions
+	return MainWithErrorCode()
+}
+
+var internalOS string
+var globalEngine workflow.Engine
+var globalConfiguration configuration.Configuration
+var globalContext context.Context
+var helpProvided bool
+
+var noopLogger zerolog.Logger = zerolog.New(io.Discard)
+var globalLogger *zerolog.Logger = &noopLogger
+var scrubbedLogger logging.ScrubbingLogWriter
+var interactionId = instrumentation.AssembleUrnFromUUID(uuid.NewString())
+
+const (
+	unknownCommandMessage     string = "unknown command"
+	disable_analytics_flag    string = "DISABLE_ANALYTICS"
+	debug_level_flag          string = "log-level"
+	integrationNameFlag       string = "integration-name"
+	maxNetworkRequestAttempts string = "max-attempts"
+	teardownTimeout                  = 5 * time.Second
+)
+
+type JsonErrorStruct struct {
+	Ok       bool   `json:"ok"`
+	ErrorMsg string `json:"error"`
+	Path     string `json:"path"`
+}
+
+type HandleError int
+
+const (
+	handleErrorFallbackToLegacyCLI HandleError = iota
+	handleErrorShowHelp            HandleError = iota
+	handleErrorUnhandled           HandleError = iota
+)
+
+// Initialize the given configuration with CLI specific aspects
+func initApplicationConfiguration(config configuration.Configuration) {
+	config.AddAlternativeKeys(configuration.AUTHENTICATION_TOKEN, []string{"snyk_cfg_api", "api"})
+	config.AddAlternativeKeys(configuration.AUTHENTICATION_BEARER_TOKEN, []string{"snyk_docker_token"})
+	config.AddAlternativeKeys(configuration.API_URL, []string{"endpoint"})
+	config.AddAlternativeKeys(configuration.ADD_TRUSTED_CA_FILE, []string{"NODE_EXTRA_CA_CERTS"})
+	config.AddAlternativeKeys(configuration.ANALYTICS_DISABLED, []string{strings.ToLower(constants.SNYK_ANALYTICS_DISABLED_ENV), "snyk_cfg_disable_analytics", "disable-analytics", "disable_analytics"})
+	config.AddAlternativeKeys(configuration.ORGANIZATION, []string{"snyk_cfg_org"})
+	config.AddAlternativeKeys(configuration.PREVIEW_FEATURES_ENABLED, []string{"snyk_preview"})
+	config.AddAlternativeKeys(configuration.LOG_LEVEL, []string{debug_level_flag})
+	config.AddAlternativeKeys(configuration.INTEGRATION_NAME, []string{integrationNameFlag})
+	config.AddAlternativeKeys(middleware.ConfigurationKeyRequestAttempts, []string{"snyk_max_attempts", maxNetworkRequestAttempts})
+	config.AddAlternativeKeys(cliv2.ConfigKeyRequestConcurrency, []string{"snyk_request_concurrency"})
+}
+
+func getFullCommandString(cmd *cobra.Command) string {
+	fullCommandPath := []string{cmd.Name()}
+	fn := func(c *cobra.Command) {
+		if c.HasParent() {
+			fullCommandPath = append(fullCommandPath, c.Name())
+		}
+	}
+	cmd.VisitParents(fn)
+
+	for i, j := 0, len(fullCommandPath)-1; i < j; i, j = i+1, j-1 {
+		fullCommandPath[i], fullCommandPath[j] = fullCommandPath[j], fullCommandPath[i]
+	}
+
+	name := strings.Join(fullCommandPath, " ")
+	return name
+}
+
+func updateConfigFromParameter(config configuration.Configuration, args []string, rawArgs []string) {
+	// extract everything behind --
+	doubleDashArgs := []string{}
+	doubleDashFound := false
+	for _, v := range rawArgs {
+		if doubleDashFound {
+			doubleDashArgs = append(doubleDashArgs, v)
+		} else if v == "--" {
+			doubleDashFound = true
+		}
+	}
+	config.Set(configuration.UNKNOWN_ARGS, doubleDashArgs)
+
+	// only consider the first positional argument as input directory if it is not behind a double dash.
+	if len(args) > 0 && !utils.Contains(doubleDashArgs, args[0]) {
+		config.Set(configuration.INPUT_DIRECTORY, args)
+		config.Set(localworkflows.ConfigurationNewAuthenticationToken, args[0])
+	}
+}
+
+// main workflow
+func runCommand(cmd *cobra.Command, args []string) error {
+	// since cobra doesn't tell us if -- was found, os.Args is required in addition
+	return runMainWorkflow(globalConfiguration, cmd, args, os.Args)
+}
+
+func runMainWorkflow(config configuration.Configuration, cmd *cobra.Command, args []string, rawArgs []string) error {
+	err := config.AddFlagSet(cmd.Flags())
+	if err != nil {
+		globalLogger.Print("Failed to add flags", err)
+		return err
+	}
+
+	// init UI
+	errorUI := consoleui.WithErrorOutput(os.Stdout)
+	if output_workflow.DefaultOutputIsStructured(config) {
+		errorUI = consoleui.WithErrorOutput(os.Stderr)
+	}
+	mainUI := consoleui.New(consoleui.WithInput(os.Stdin), consoleui.WithOutput(os.Stdout), consoleui.WithProgressWriter(os.Stderr), errorUI)
+	globalEngine.SetUserInterface(mainUI)
+
+	// global handling of experimental commands
+	if config_utils.IsExperimental(cmd.Flags()) {
+		if !globalConfiguration.GetBool(configuration.FLAG_EXPERIMENTAL) {
+			return cli.NewCommandIsExperimentalError(getFullCommandString(cmd))
+		}
+	}
+
+	updateConfigFromParameter(config, args, rawArgs)
+
+	name := getFullCommandString(cmd)
+	globalLogger.Print("Running ", name)
+	globalEngine.GetAnalytics().SetCommand(name)
+
+	err = runWorkflowAndProcessData(globalContext, globalEngine, globalLogger, name)
+
+	return err
+}
+
+func runWorkflowAndProcessData(ctx context.Context, engine workflow.Engine, logger *zerolog.Logger, name string) error {
+	ic := engine.GetAnalytics().GetInstrumentation()
+
+	output, err := engine.Invoke(workflow.NewWorkflowIdentifier(name), workflow.WithContext(ctx), workflow.WithInstrumentationCollector(ic))
+	if err != nil {
+		logger.Print("Failed to execute the command! ", err)
+		return err
+	}
+
+	outputFiltered, err := engine.Invoke(localworkflows.WORKFLOWID_FILTER_FINDINGS, workflow.WithContext(ctx), workflow.WithInput(output), workflow.WithInstrumentationCollector(ic))
+	if err != nil {
+		logger.Err(err).Msg(err.Error())
+		return err
+	}
+
+	_, err = engine.Invoke(localworkflows.WORKFLOWID_OUTPUT_WORKFLOW, workflow.WithContext(ctx), workflow.WithInput(outputFiltered), workflow.WithInstrumentationCollector(ic))
+	if err == nil {
+		err = getErrorFromWorkFlowData(engine, outputFiltered)
+	}
+	return err
+}
+
+func help(_ *cobra.Command, _ []string) error {
+	helpProvided = true
+	args := utils.RemoveSimilar(os.Args[1:], "--") // remove all double dash arguments to avoid issues with the help command
+	args = append(args, "--help")
+	return defaultCmd(args)
+}
+
+func defaultCmd(args []string) error {
+	inputDirectory := cliv2.DetermineInputDirectory(args)
+	if len(inputDirectory) > 0 {
+		globalConfiguration.Set(configuration.INPUT_DIRECTORY, inputDirectory)
+	}
+
+	// prepare the invocation of the legacy CLI by
+	// * enabling stdio
+	// * by specifying the raw cmd args for it
+	globalConfiguration.Set(configuration.WORKFLOW_USE_STDIO, true)
+	globalConfiguration.Set(configuration.RAW_CMD_ARGS, args)
+	_, err := globalEngine.Invoke(basic_workflows.WORKFLOWID_LEGACY_CLI)
+	return err
+}
+
+func runTestCommandWithSarifEqualJson(cmd *cobra.Command, args []string, templateFiles []string) error {
+	// ensure legacy behavior, where sarif and json can be used interchangeably
+	globalConfiguration.AddAlternativeKeys(output_workflow.OUTPUT_CONFIG_KEY_SARIF, []string{output_workflow.OUTPUT_CONFIG_KEY_JSON})
+
+	fileWriters := []output_workflow.FileWriter{
+		{
+			NameConfigKey:     output_workflow.OUTPUT_CONFIG_KEY_SARIF_FILE,
+			MimeType:          output_workflow.SARIF_MIME_TYPE,
+			TemplateFiles:     templateFiles,
+			WriteEmptyContent: true,
+		},
+		{
+			NameConfigKey:     output_workflow.OUTPUT_CONFIG_KEY_JSON_FILE,
+			MimeType:          output_workflow.SARIF_MIME_TYPE,
+			TemplateFiles:     templateFiles,
+			WriteEmptyContent: false,
+		},
+	}
+	globalConfiguration.Set(output_workflow.OUTPUT_CONFIG_KEY_FILE_WRITERS, fileWriters)
+
+	// ensure that json is translated to sarif for the default writer as well
+	defaultWriterLookup := map[string]string{
+		output_workflow.JSON_MIME_TYPE: output_workflow.SARIF_MIME_TYPE,
+	}
+	globalConfiguration.Set(output_workflow.OUTPUT_CONFIG_KEY_DEFAULT_WRITER_LUT, defaultWriterLookup)
+
+	return runCommand(cmd, args)
+}
+
+func runCodeTestCommand(cmd *cobra.Command, args []string) error {
+	return runTestCommandWithSarifEqualJson(cmd, args, output_workflow.ApplicationSarifTemplates)
+}
+
+func runSecretsTestCommand(cmd *cobra.Command, args []string) error {
+	return runTestCommandWithSarifEqualJson(cmd, args, output_workflow.ApplicationSarifTemplatesUfm)
+}
+
+func runAuthCommand(cmd *cobra.Command, args []string) error {
+	err := runCommand(cmd, args)
+
+	reloadError := globalConfiguration.ReloadConfig()
+	if reloadError != nil {
+		globalLogger.Err(reloadError).Msg("Failed to reload the configuration after authentication.")
+	}
+
+	return err
+}
+
+func getGlobalFLags() *pflag.FlagSet {
+	globalConfigurationOptions := workflow.GetGlobalConfiguration()
+	globalFLags := workflow.FlagsetFromConfigurationOptions(globalConfigurationOptions)
+	globalFLags.Bool(basic_workflows.PROXY_NOAUTH, false, "")
+	globalFLags.Bool(disable_analytics_flag, false, "")
+	globalFLags.String(debug_level_flag, "debug", "")
+	globalFLags.String(integrationNameFlag, "", "")
+	globalFLags.Int(maxNetworkRequestAttempts, -1, "Maximum total network attempts, including the initial request (minimum: 1)")
+	return globalFLags
+}
+
+func emptyCommandFunction(_ *cobra.Command, _ []string) error {
+	return fmt.Errorf("%s", unknownCommandMessage)
+}
+
+func createCommandsForWorkflows(rootCommand *cobra.Command, engine workflow.Engine) {
+	workflowIdList := engine.GetWorkflows()
+	commandMap := make(map[string]*cobra.Command)
+	for i := range workflowIdList {
+		currentId := workflowIdList[i]
+		currentCommandString := workflow.GetCommandFromWorkflowIdentifier(currentId)
+
+		workflowEntry, _ := engine.GetWorkflow(currentId)
+
+		workflowOptions := workflowEntry.GetConfigurationOptions()
+		flagset := workflow.FlagsetFromConfigurationOptions(workflowOptions)
+
+		workflowParts := strings.Split(currentCommandString, " ")
+
+		parentCommand := rootCommand
+		var currentParts string
+		for level := 0; level < len(workflowParts); level++ {
+			subCmdName := workflowParts[level]
+			currentParts += " " + subCmdName
+			subCmd := commandMap[currentParts]
+			if subCmd == nil {
+				subCmd = &cobra.Command{
+					Use:                subCmdName,
+					Hidden:             true,
+					RunE:               emptyCommandFunction, // ensure to trigger the fallback case
+					DisableFlagParsing: true,                 // disable flag parsing to allow arbitrary flags for commands that will trigger the fallback
+				}
+				parentCommand.AddCommand(subCmd)
+				commandMap[currentParts] = subCmd
+			}
+			parentCommand = subCmd
+		}
+
+		// last parentCommand is the last level of the command, e.g. in the snyk iac capture workflow, it would be 'capture'
+		// we add flags and command line only to that one
+		if flagset != nil {
+			parentCommand.Flags().AddFlagSet(flagset)
+		}
+		parentCommand.RunE = runCommand
+		parentCommand.Hidden = !workflowEntry.IsVisible()
+		parentCommand.DisableFlagParsing = false
+
+		// special case for snyk code test
+		switch currentCommandString {
+		case "code test":
+			// to preserve backwards compatibility we will need to relax flag validation
+			parentCommand.FParseErrWhitelist.UnknownFlags = true
+
+			// use the special run command to ensure that the non-standard behavior of the command can be kept
+			parentCommand.RunE = runCodeTestCommand
+		case "secrets test":
+			// use the special run command to ensure that the non-standard behavior of the command can be kept
+			parentCommand.RunE = runSecretsTestCommand
+		case "test", "monitor":
+			legacy.SetupTestMonitorCommand(parentCommand)
+		case "auth":
+			parentCommand.RunE = runAuthCommand
+		case "agent-scan":
+			// to preserve backwards compatibility we will need to relax flag validation
+			parentCommand.FParseErrWhitelist.UnknownFlags = true
+			parentCommand.Aliases = []string{"mcp-scan"}
+		}
+	}
+}
+
+func prepareRootCommand() *cobra.Command {
+	rootCommand := cobra.Command{
+		Use: "snyk",
+		RunE: func(cmd *cobra.Command, _ []string) error {
+			return defaultCmd(os.Args[1:])
+		},
+	}
+
+	// help for all commands is handled by the legacy cli
+	// TODO: discuss how to move help to extensions
+	helpCommand := cobra.Command{
+		Use:  "help",
+		RunE: help,
+	}
+
+	// some static/global cobra configuration
+	rootCommand.CompletionOptions.DisableDefaultCmd = true
+	rootCommand.SilenceErrors = true
+	rootCommand.SilenceUsage = true
+	rootCommand.FParseErrWhitelist.UnknownFlags = true
+
+	// ensure that help and usage information comes from the legacy cli instead of cobra's default help
+	rootCommand.SetHelpFunc(func(c *cobra.Command, args []string) { _ = help(c, args) })
+	rootCommand.SetHelpCommand(&helpCommand)
+	rootCommand.PersistentFlags().AddFlagSet(getGlobalFLags())
+
+	return &rootCommand
+}
+
+func handleError(err error) HandleError {
+	resultError := handleErrorUnhandled
+	preCondition := err != nil && !helpProvided
+
+	// Cases:
+	// - error from extension -> ignore
+	// - error unknown command -> fallback
+	// - error known command but unknown flag -> help
+
+	if preCondition {
+		errString := err.Error()
+		flagError := strings.Contains(errString, "unknown flag") ||
+			strings.Contains(errString, "flag needs") ||
+			strings.Contains(errString, "invalid argument")
+		commandError := strings.Contains(errString, unknownCommandMessage)
+
+		// filter for known cobra errors, since cobra errors shall trigger a fallback, but not others.
+		if commandError {
+			resultError = handleErrorFallbackToLegacyCLI
+		} else if flagError {
+			// handle flag errors explicitly since we need to delegate the help to the legacy CLI. This includes disabling the cobra default help/usage
+			resultError = handleErrorShowHelp
+		}
+	}
+
+	return resultError
+}
+
+func displayError(err error, userInterface ui.UserInterface, config configuration.Configuration, ctx context.Context) {
+	if err != nil {
+		_, isExitError := err.(*exec.ExitError)
+		_, isErrorWithCode := err.(*cli_errors.ErrorWithExitCode)
+		if isExitError || isErrorWithCode || errorHasBeenShown(err) {
+			return
+		}
+
+		if config.GetBool(output_workflow.OUTPUT_CONFIG_KEY_JSON) {
+			message := getErrorMessage(err)
+
+			jsonError := JsonErrorStruct{
+				Ok:       false,
+				ErrorMsg: message,
+				Path:     globalConfiguration.GetString(configuration.INPUT_DIRECTORY),
+			}
+
+			jsonErrorBuffer, _ := json.MarshalIndent(jsonError, "", "  ")
+			_ = userInterface.OutputError(fmt.Errorf("%s", jsonErrorBuffer))
+		} else {
+			uiError := userInterface.OutputError(err, ui.WithContext(ctx))
+			if uiError != nil {
+				globalLogger.Err(uiError).Msg("ui failed to show error")
+			}
+		}
+	}
+}
+
+func initExtensions(engine workflow.Engine, config configuration.Configuration) {
+	engine.AddExtensionInitializer(basic_workflows.Init)
+	engine.AddExtensionInitializer(osflows.Init)
+	engine.AddExtensionInitializer(iac.Init)
+	engine.AddExtensionInitializer(sbom.Init)
+	engine.AddExtensionInitializer(aibom.Init)
+	engine.AddExtensionInitializer(redteam.Init)
+	engine.AddExtensionInitializer(depgraph.Init)
+	engine.AddExtensionInitializer(capture.Init)
+	engine.AddExtensionInitializer(iacrules.Init)
+	engine.AddExtensionInitializer(snykls.Init)
+	engine.AddExtensionInitializer(mcp.Init)
+	engine.AddExtensionInitializer(container.Init)
+	engine.AddExtensionInitializer(code.Init)
+	engine.AddExtensionInitializer(workflows.InitConnectivityCheckWorkflow)
+	engine.AddExtensionInitializer(ignoreworkflow.InitIgnoreWorkflows)
+	engine.AddExtensionInitializer(agentscan.Init)
+	engine.AddExtensionInitializer(secrets.Init)
+
+	// Register additional extensions injected via Run(WithAdditionalExtensions(...))
+	for _, ext := range additionalExts {
+		engine.AddExtensionInitializer(ext)
+	}
+
+	if config.GetBool(configuration.PREVIEW_FEATURES_ENABLED) {
+		config.Set("INTERNAL_USE_UFM_PRESENTER", true)
+	}
+}
+
+// tearDown handles sending analytics and instrumentation
+// It is used both for normal exit and signal-triggered exit
+func tearDown(err error, errorList []error, startTime time.Time, ua networking.UserAgentInfo, cliAnalytics analytics.Analytics, networkAccess networking.NetworkAccess) int {
+	// Create a context with timeout for teardown operations to ensure we don't hang indefinitely
+	teardownCtx, cancel := context.WithTimeout(context.Background(), teardownTimeout)
+	defer cancel()
+
+	outputError := err
+	allErrors := errorList
+
+	if err != nil {
+		allErrors, outputError = processError(err, errorList)
+
+		for _, tempError := range allErrors {
+			if tempError != nil {
+				cliAnalytics.AddError(tempError)
+			}
+		}
+	}
+
+	exitCode := cliv2.DeriveExitCode(outputError)
+	globalLogger.Printf("Deriving Exit Code %d (cause: %v)", exitCode, outputError)
+
+	displayError(outputError, globalEngine.GetUserInterface(), globalConfiguration, globalContext)
+
+	updateInstrumentationDataBeforeSending(cliAnalytics, startTime, ua, exitCode)
+
+	if !globalConfiguration.GetBool(configuration.ANALYTICS_DISABLED) {
+		sendAnalytics(teardownCtx, cliAnalytics, globalLogger)
+	}
+	sendInstrumentation(teardownCtx, globalEngine, cliAnalytics.GetInstrumentation(), globalLogger)
+
+	// cleanup resources in use
+	// WARNING: deferred actions will execute AFTER cleanup; only defer if not impacted by this
+	if _, cleanupErr := globalEngine.Invoke(basic_workflows.WORKFLOWID_GLOBAL_CLEANUP, workflow.WithContext(teardownCtx)); cleanupErr != nil {
+		globalLogger.Printf("Failed to cleanup %v", cleanupErr)
+	}
+
+	if globalConfiguration.GetBool(configuration.DEBUG) {
+		writeLogFooter(exitCode, allErrors, globalConfiguration, networkAccess)
+	}
+
+	return exitCode
+}
+
+func MainWithErrorCode() int {
+	initDebugBuild()
+
+	errorList := []error{}
+	errorListMutex := sync.Mutex{}
+	var finalExitCode int
+
+	// preparing the possibility to tearDown from different threads while ensure it is only called once
+	var tearDownOnce sync.Once
+
+	// init context
+	ctx := context.Background()
+	ctx = context.WithValue(ctx, networking.InteractionIdKey, instrumentation.AssembleUrnFromUUID(interactionId))
+	globalContext = ctx
+
+	startTime := time.Now()
+	var err error
+	rInfo := runtimeinfo.New(runtimeinfo.WithName("snyk-cli"), runtimeinfo.WithVersion(cliv2.GetFullVersion()))
+
+	rootCommand := prepareRootCommand()
+	// omit the first arg which is always `snyk`
+	_ = rootCommand.ParseFlags(os.Args[1:])
+
+	// create engine
+	globalConfiguration = configuration.NewWithOpts(
+		configuration.WithFiles("snyk"),
+		configuration.WithSupportedEnvVars("NODE_EXTRA_CA_CERTS"),
+		configuration.WithSupportedEnvVarPrefixes("snyk_", "internal_", "test_"),
+		configuration.WithCachingEnabled(configuration.NoCacheExpiration),
+	)
+	err = globalConfiguration.AddFlagSet(rootCommand.LocalFlags())
+	if err != nil {
+		fmt.Fprintln(os.Stderr, "Failed to add flags to root command", err)
+	}
+
+	// ensure to init configuration before using it
+	initApplicationConfiguration(globalConfiguration)
+
+	debugEnabled := globalConfiguration.GetBool(configuration.DEBUG)
+
+	globalLogger, scrubbedLogger = initDebugLogger(globalConfiguration)
+	globalEngine = app.CreateAppEngineWithOptions(app.WithZeroLogger(globalLogger), app.WithConfiguration(globalConfiguration), app.WithRuntimeInfo(rInfo))
+
+	globalConfiguration.AddDefaultValue(configuration.FF_OAUTH_AUTH_FLOW_ENABLED, defaultOAuthFF(globalConfiguration))
+	globalConfiguration.AddDefaultValue(configuration.FF_TRANSFORMATION_WORKFLOW, configuration.StandardDefaultValueFunction(true))
+
+	if noProxyAuth := globalConfiguration.GetBool(basic_workflows.PROXY_NOAUTH); noProxyAuth {
+		globalConfiguration.Set(configuration.PROXY_AUTHENTICATION_MECHANISM, httpauth.StringFromAuthenticationMechanism(httpauth.NoAuth))
+	}
+
+	// init NetworkAccess
+	ua := networking.UserAgent(networking.UaWithConfig(globalConfiguration), networking.UaWithRuntimeInfo(rInfo), networking.UaWithOS(internalOS))
+	networkAccess := globalEngine.GetNetworkAccess()
+	networkAccess.AddErrorHandler(func(err error, ctx context.Context) error {
+		if err == nil {
+			return nil
+		}
+		errorListMutex.Lock()
+		defer errorListMutex.Unlock()
+
+		errorList = append(errorList, err)
+		return err
+	})
+	networkAccess.AddHeaderField("x-snyk-cli-version", cliv2.GetFullVersion())
+	networkAccess.AddHeaderField("snyk-interaction-id", instrumentation.AssembleUrnFromUUID(interactionId))
+	networkAccess.AddHeaderField(
+		"User-Agent",
+		ua.String(),
+	)
+	network_utils.AddSnykRequestId(networkAccess)
+
+	if debugEnabled {
+		writeLogHeader(globalConfiguration, networkAccess)
+	}
+
+	// initialize the extensions -> they register themselves at the engine
+	initExtensions(globalEngine, globalConfiguration)
+
+	// init engine
+	err = globalEngine.Init()
+
+	// We want to scrub the debug log of sensitive information. Since we have a list of commands we know can occur, we can intersect that with arguments we don't recognize, and automatically scrub all those from the logs.
+	if debugEnabled {
+		knownTerms, _ := instrumentation.GetKnownCommandsAndFlags(globalEngine)
+		knownTerms = append(knownTerms, globalConfiguration.GetString(configuration.API_URL), globalConfiguration.GetString(configuration.ORGANIZATION), globalConfiguration.GetString(configuration.ORGANIZATION_SLUG))
+		termsToRedact := cliv2utils.GetUnknownParameters(os.Args[1:], os.Environ(), knownTerms)
+		scrubbedLogger.AddTermsToReplace(termsToRedact)
+	}
+
+	if err != nil {
+		globalLogger.Print("Failed to init Workflow Engine!", err)
+		return constants.SNYK_EXIT_CODE_ERROR
+	}
+
+	// add output flags as persistent flags
+	outputWorkflow, _ := globalEngine.GetWorkflow(localworkflows.WORKFLOWID_OUTPUT_WORKFLOW)
+	outputFlags := workflow.FlagsetFromConfigurationOptions(outputWorkflow.GetConfigurationOptions())
+	rootCommand.PersistentFlags().AddFlagSet(outputFlags)
+
+	// add workflows as commands
+	createCommandsForWorkflows(rootCommand, globalEngine)
+
+	// init Analytics
+	cliAnalytics := globalEngine.GetAnalytics()
+	cliAnalytics.SetVersion(cliv2.GetFullVersion())
+	cliAnalytics.SetCmdArguments(os.Args)
+	cliAnalytics.SetOperatingSystem(internalOS)
+	cliAnalytics.GetInstrumentation().SetUserAgent(ua)
+	cliAnalytics.GetInstrumentation().SetInteractionId(instrumentation.AssembleUrnFromUUID(interactionId))
+	cliAnalytics.GetInstrumentation().SetCategory(instrumentation.DetermineCategory(os.Args, globalEngine))
+	cliAnalytics.GetInstrumentation().SetStage(instrumentation.DetermineStage(cliAnalytics.IsCiEnvironment()))
+	cliAnalytics.GetInstrumentation().SetStatus(analytics.Success)
+
+	setTimeout(globalConfiguration, func() {
+		tearDownOnce.Do(func() {
+			errorListMutex.Lock()
+			errorListCopy := append([]error{}, errorList...)
+			errorListMutex.Unlock()
+
+			exitCode := tearDown(context.DeadlineExceeded, errorListCopy, startTime, ua, cliAnalytics, networkAccess)
+			os.Exit(exitCode)
+		})
+	})
+
+	// run the extensible cli
+	err = rootCommand.Execute()
+
+	// fallback to the legacy cli or show help
+	handleErrorResult := handleError(err)
+	switch handleErrorResult {
+	case handleErrorFallbackToLegacyCLI:
+		// when falling back to TS cli, make sure the
+		_ = rootCommand.ParseFlags(os.Args)
+		if err := globalConfiguration.AddFlagSet(rootCommand.LocalFlags()); err != nil {
+			globalLogger.Printf("failed to add flagset: %v", err)
+		}
+
+		globalLogger.Printf("Using Legacy CLI to serve the command. (reason: %v)", err)
+		err = defaultCmd(os.Args[1:])
+	case handleErrorShowHelp:
+		err = help(nil, []string{})
+	case handleErrorUnhandled:
+		// ignore
+	}
+
+	tearDownOnce.Do(func() {
+		errorListMutex.Lock()
+		errorListCopy := append([]error{}, errorList...)
+		errorListMutex.Unlock()
+
+		finalExitCode = tearDown(err, errorListCopy, startTime, ua, cliAnalytics, networkAccess)
+	})
+
+	return finalExitCode
+}
+
+func processError(err error, errorList []error) ([]error, error) {
+	// ensure to use generic fallback error catalog error if no other is available
+	resultError := decorateError(err)
+	resultErrorList := errorList
+
+	// filter legacycli terminate errors since it is only used for internal purposes
+	if exitErr, isExitError := resultError.(*exec.ExitError); isExitError && exitErr.ExitCode() == constants.SNYK_EXIT_CODE_TS_CLI_TERMINATED {
+		resultError = nil
+	}
+
+	// ensure to have all errors in the list for analytics, determining the most relevant error to surface to the user ...
+	if resultError != nil {
+		resultErrorList = append([]error{resultError}, resultErrorList...)
+	}
+
+	// determine the most relevant error to surface to the user and derive the exit code from
+	resultError = cli_errors.FindMostRelevantError(resultErrorList)
+
+	// ensure to apply exit code mapping based on errors
+	if exitCode := mapErrorToExitCode(resultError); exitCode != unsetExitCode {
+		resultError = createErrorWithExitCode(exitCode, resultError)
+	}
+	return resultErrorList, resultError
+}
+
+func setTimeout(config configuration.Configuration, onTimeout func()) {
+	timeout := config.GetInt(configuration.TIMEOUT)
+	if timeout == 0 {
+		return
+	}
+	globalLogger.Printf("Command timeout set for %d seconds", timeout)
+	go func() {
+		const gracePeriodForSubProcesses = 3
+		<-time.After(time.Duration(timeout+gracePeriodForSubProcesses) * time.Second)
+		onTimeout()
+	}()
+}
