+++ cliv2/pkg/runner/runner_test.go
@@ -1,4 +1,4 @@
-package main
+package runner
 
 import (
 	"encoding/json"
